Handle positions correctly for present but empty fields
It is fine for a field to be present in a document, but have no
values - think of tokenising an empty string.

This update makes the necessary changes to handle this edgecase
correctly, including modifying the schema to allow entries in
the position_doc table that record the absence of positions.

FossilOrigin-Name: e94df37cc150b69e11b0613c9ca9161ae5f347471ce6a0c0e5a739b84ca200be