Handle no-op edgecase for clustering that cannot change
A clustering with a single cluster selected and target_clusters
not set, or set to 1 cannot change. This is still valid though,
so now this is a no-op and the initial clustering is immediately
returned.

This closes #64.

FossilOrigin-Name: 5099f9b2401c96d185d2f8a5990ba5977d2050a6afde8db95778e0653a3bc40b